// src/markdown/types.ts
export interface PageNavigation {
  current: PageNavigationData | null;
  parent: PageNavigationData | null;
  next: PageNavigationData | null;
  previous: PageNavigationData | null;
}

export interface PageNavigationData {
  title: string | null;
  url: string | null;
}